Add support for more CPS formats * rename QSound driver to CPS * add support for CPS3 sequence/instr/samp formats * add support for CPS1 sequences * add CPS3 decryption to MAMELoader
