+++ src/main/loaders/MAMELoader.h
@@ -13,8 +13,8 @@ enum LoadMethod {
 
 using namespace std;
 
-typedef struct _MAMERomGroupEntry {
-  _MAMERomGroupEntry() : file(NULL) { }
+struct MAMERomGroup {
+  MAMERomGroup() : file(NULL) { }
   template<class T>
   bool GetAttribute(const std::string &attrName, T *out) {
     string strValue = attributes[attrName];
@@ -32,38 +32,37 @@ typedef struct _MAMERomGroupEntry {
   std::map<const std::string, std::string> attributes;
   std::list<std::string> roms;
   VirtFile *file;
-} MAMERomGroupEntry;
+};
 
-typedef struct _MAMEGameEntry {
-  _MAMEGameEntry() { }
-  MAMERomGroupEntry *GetRomGroupOfType(const std::string &strType);
+struct MAMEGame {
+  MAMEGame() { }
+  MAMERomGroup *GetRomGroupOfType(const std::string &strType);
 
   std::string name;
   std::string format;
-  float fmt_version;
   std::string fmt_version_str;
   //map<const std::string, const std::string> attributes;
-  std::list<MAMERomGroupEntry> romgroupentries;
-} MAMEGameEntry;
+  std::list<MAMERomGroup> romgroupentries;
+};
 
-typedef std::map<std::string, MAMEGameEntry *> GameMap;
+typedef std::map<std::string, MAMEGame *> GameMap;
 
 
 class MAMELoader:
     public VGMLoader {
- public:
+public:
   MAMELoader();
   ~MAMELoader();
   virtual PostLoadCommand Apply(RawFile *theFile);
- private:
-  VirtFile *LoadRomGroup(MAMERomGroupEntry *romgroupentry, const std::string &format, unzFile &cur_file);
+private:
+  VirtFile *LoadRomGroup(MAMERomGroup *romgroupentry, const std::string &format, unzFile &cur_file);
   void DeleteBuffers(std::list<std::pair<uint8_t *, uint32_t>> &buffers);
- private:
+private:
   int LoadXML();
-  MAMEGameEntry *LoadGameEntry(TiXmlElement *gameElmt);
-  int LoadRomGroupEntry(TiXmlElement *romgroupElmt, MAMEGameEntry *gameentry);
+  MAMEGame *LoadGameEntry(TiXmlElement *gameElmt);
+  int LoadRomGroupEntry(TiXmlElement *romgroupElmt, MAMEGame *gameentry);
 
- private:
+private:
   GameMap gamemap;
   bool bLoadedXml;
 };
